Technical duty

+++ app/src/main/java/sleepless_nights/location_alarm/alarm/view_models/AlarmViewModel.java
@@ -38,9 +38,14 @@ public class AlarmViewModel extends AndroidViewModel {
     }
 
     @Nullable
-    public Alarm getAlarmLiveDataById(int id) {
+    public Alarm getAlarmLiveDataById(long id) {
         return alarmRepository.getAlarmById(id);
     }
+
+    /*
+    * TODO #4 make not read-only repository operations not void
+    *  we should return some structure which can accept onSuccess and onFailure callbacks
+    */
 
     public void createAlarm(String name, String address, boolean isActive,
                             double latitude, double longitude, float radius) {
